lxd_vm - use import_role for zfsdocker
The issue is that zfsdocker_name is passed to lxd_vm_docker, which uses it as
part of the zfsdocker_zvol variable. This variable is then used later in the
lxd_vm role.

However! By then the zfsdocker_name variable is no longer in scope, since we
used include_role (which we did in an effort to avoid leaking variables in the
first place).

Does this sound unnecessarily complicated? That's because it is. You can thank
ansible for the extra gray hair and unwanted job security.