postgres refactor - prefer shell=1 for check_output
While there is something to be said for interpolating string input and simply
passing it to a shell, it's mostly relevant when accepting input over the
internet or when running with elevated privileges.

Using shell=True (or =1) is easier. If one wants to use the list syntax for
subprocess.run() then the shlex module should be used for parsing the string
instead of simply split().